src/layers/shared/requirements.txt
aws-xray-sdk==2.14.0
nanoid==2.0.0
pyjwt[crypto]==2.12.1
pynacl==1.5.0
requests==2.32.5
